Advertise public heat endpoint for wait conditions
Wait conditions are commonly used by instances to signal various events back to
heat. These instances are unlikely to have access to the internal API
endpoints. OpenStack-Ansible had a similar issue[1] back in juno and changed to
use the public endpoint[2]. The code has now moved but the default is still in
place[3].

[1] https://bugs.launchpad.net/openstack-ansible/+bug/1459414
[2] https://review.openstack.org/#/c/186221/
[3] https://github.com/openstack/openstack-ansible-os_heat/blob/b1721a7460ba816afabd9bded6e5c79c635bca3a/defaults/main.yml#L48

Closes-Bug: #1688331