Reinstate [DEFAULT] firewall_driver option
RDO packages a distribution configuration file
/usr/share/nova/nova-dist.conf which contains the following setting:

firewall_driver = nova.virt.libvirt.firewall.IptablesFirewallDriver

This causes the nova-compute-ironic service to fail to start as the
nova ironic virt driver attempts and fails to create a firewall driver
using this class.

This change reinstates the explicit setting of the [DEFAULT]
firewall_driver option to the noop driver which resolves this issue.
This comes at the cost of a WARN log message due to the option being
deprecated (see
https://github.com/openstack/kolla-ansible/commit/6d831db687a35a48a41ee581b979fb43350e0c72).

Change-Id: I41bd9d0671118ff256e7ada766e8653bb4b2b376
Closes-Bug: #1701564