Add support for CephFS Native Manila backend
One change that affects all manila backends is the configuration of the
enabled_share_protocols. This defaults to "NFS, CIFS", which I have
maintained for the existing backends. When it comes to the CephFS native
driver we need to add CEPHFS into the list.

Generally this has copied the pattern used in glance to support
connecting to an external ceph system, mixed with the existing way
manila backends are enabled.

Change-Id: If1c4307d47c3607a97c8b32d3f06fc64e80e94b4