Factor out nova-cell role
Splitting this role out from the Nova role allows a more granular
approach to deploying and configuring Nova services. It paves the
way for having a role which can be used to deploy multiple Nova cells.

This patch does not support using a single cell controller for managing
more than one cell. Support for sharing a cell controller will be added
in a future patch.

This patch should be backwards compatible and is tested by existing CI
jobs. A Future patch will add a multi-cell test.

Partially Implements: blueprint support-nova-cells
Co-Authored-By: Mark Goddard <mark@stackhpc.com>
Change-Id: I810aad7d49db3f5a7fd9a2f0f746fd912fe03917