Minimal support for deploying multiple instances of MariaDB
This change adds minimal support for deploying multiple
instances of MariaDB which are not on the same node.

This change is useful for deploying one database engine instance
per service, and for the Cells v2 work.

The container name, docker volume, configuration and log files
are all namespaced.

This change should maintain the default behaviour.

Partially Implements: blueprint support-nova-cells
Depends-On: https://review.opendev.org/#/c/675386
Change-Id: I7c30a7c3bace22d1e70e395e8a86c572f8fe5ee4