Extract useful information from HAProxy logs
The HAProxy logs contain detailed information about HTTP and TCP
connections which are not available from the HAProxy exporter. For
example, the information from these logs can be used to measure the
response time for an HTTP request as a function of status code or
'tail latency' as a function of service and backend server.

Change-Id: I716e0d6d42a26c53aae3519db5b26b05ed60b253
(cherry picked from commit 2abd7d44a52a958f99ca7429db9252683dc0a33f)