WIP: Staged start for Monasca Python consumers
Whilst Kafka auto topic creation is on by default, and consumers
will automatically create topics that they subscribe to, we appear
to hit a race where when multiple consumers of the same type start
at around the same time, some of them will return an error relating
to a missing topic. This change attempts a staged start in
multi-node systems to work around this issue.

The mechanism for waiting for the service to start could be improved,
but is hopefully enough.

Note that although this could be considered a bug, it isn't an issue
outside of CI because the consumer containers will just restart.

Change-Id: I3e5ed7752f6d6e791dad2a6aec27e380d2b1da9e