Fix prechecks with Docker 20.10.0 and py27 job
The 'prechecks : Checking Docker version' task previously failed with
Docker 20.10.0. The regex used to parse the version was returning
0.10.0, which is not above the minimum. The previous version of 19.x
would have been parsed as 9.x, which is above the minimum.

This change fixes the issue by matching the beginning and end of the
version using \b.

For Train and below additionally blacklist bandit 1.6.3 as
it was improperly released as a universal wheel. [1]
And also drop lower-constraints job as agreed during the PTG.
[2] It started to fail.

[1] https://github.com/PyCQA/bandit/issues/663
[2] http://lists.openstack.org/pipermail/openstack-discuss/2020-October/018445.html

Depends-On: https://review.opendev.org/766180

Change-Id: I2a23eea7effb5b9a5e73361bcd48bd2e16d1569c
Closes-Bug: 1907436
Co-Authored-By: Radosław Piliszek <radoslaw.piliszek@gmail.com>