libvirt: make it possible to run libvirt on the host
In some cases it may be desirable to run the libvirt daemon on the host.
For example, when mixing host and container OS distributions or
versions.

This change implements changes to the baremetal role necessary to
disable the nova_libvirt container. In this case we should not remove
libvirt packages from the host, nor should we remove the AppArmor
profile for libvirt on Ubuntu hosts.

Needed-By: https://review.opendev.org/c/openstack/kolla-ansible/+/825357

Change-Id: I8dbe805cea66bd04374b36a4e8876da9b05b2045
(cherry picked from commit efd3335f6d8add0a4bf15f98bb744e990352bbfb)