Don't assume /etc/hosts exists
The baremetal role adds entries to /etc/hosts. If the file does not
exist, it will fail:

  TASK [baremetal : Ensure localhost in /etc/hosts] ****************************************
  fatal: [compute2]: FAILED! => {"changed": false, "msg": "Destination /etc/hosts does not exist !", "rc": 257}

Though surprising, this is possible, and we should not assume that
/etc/hosts exists.

This change modifies the above task to create the file if it does not
exist.

Closes-Bug: #1982837
Change-Id: Ic419ee13a10b3fa1b890cfec23bab0a67489eb96
(cherry picked from commit 25020e469f4a6f913ebaf21a3dc812d1fcbfa53a)