Adds api_hostnames to baremetal role
This allows you to customise the entries that are added to /etc/hosts.
The names in here determine the name used to register the compute service,
since nova uses socket.getfqdn()[1]. The fact `nodename` seems
to be looked up as follows:

- ansible calls platform.node() from standard library
- platform.node seems to use os.uname() from standard library
- os.uname seems to do a uname syscall
- uname syscall uses data in /proc/sys/kernel/hostname

Generally this should be set to shortname[2]. The nodename fact will
therefore in general return the shortname. Instead, we should explictly
set the FQDN so that we can reliably control what socket.getfqdn()
returns.

[1] https://github.com/openstack/nova/commit/be3c54ee0d886c38f125727beba3a72b75bb32e6
[2] https://unix.stackexchange.com/a/239950

Change-Id: I1ffc7ac4cbd35bbf1d432b76ccd18298dcd57315
(cherry picked from commit 111bc6475304d36abe3fc17287122905e1eba014)