WIP: Reduce agent restart storm
This is an attempt to minimise l3 downtime. The theory is that
restarting many agents at the same time is overloading neutron server.
This creates a backlog of messages in the neutron RPC queue which can
take some time to clear. The hope is that by increasing the period of
time over which the agents restart (by limiting to one host at a time)
will help spread the load such that neutron can keep up with processing
its message queue.

Change-Id: I8a0aef80a7a3cb25e51ab6b3f7c0138cc0ded438