Temporarily set kombu_reconnect_delay to 0.1
Temporarily set the value [oslo_messaging_rabbit] kombu_reconnect_delay
to 0.1 in all services which use RabbitMQ. Fixes the "timeout waiting
for reply..." errors seen, which cause VMs to fail building when a
RabbitMQ node is down. See bug:
https://bugs.launchpad.net/cloud-archive/+bug/1993149

This patch is temporary while the upstream patche(s) to oslo.messaging
are still in progress. It should be reverted once the bug is fixed
upstream. See:
https://review.opendev.org/c/openstack/oslo.messaging/+/866617
https://review.opendev.org/c/openstack/oslo.messaging/+/866616