Enhance TLS configuration and certificate handling in Kolla
- Updated `kolla_enable_tls_external` logic in `group_vars/all.yml` to
  ensure TLS is only enabled when external and internal VIPs are distinct.
- Refined `kolla_copy_ca_into_containers` to dynamically enable CA copy
  based on TLS configuration and certificate management settings.
- Added additional condition to `copy-certs.yml` to ensure certificates
  are copied only when external and internal VIPs differ.

These changes improve flexibility in TLS configuration and ensure more
accurate handling of certificates for both internal and external endpoints.

These defaults make more sense because they work correctly. The only side
effect is that if the user explicitly enables TLS for internal when using
two ips, it also enables it for external. However, this is more expected
than the current behavior and don't brake logic. And the user can still
explicitly disable external if needed, and we have also prechecks.

Closes-Bug: #2091083
Change-Id: Icdc923b45f916c8ca5380c1758cc1aac647db9af