Use file service discovery in Prometheus
This change switches from using static target lists to file-based
service discovery [1] to discover scrape targets.

Prometheus now watches a configuration directory for changes rather than
only loading files on startup.

The primary advantage of this is that the prometheus server container
does not need to be restarted for most configuration changes.

[1] https://prometheus.io/docs/guides/file-sd/

Change-Id: I11d35b47b0481215a84b18dfcfa27521b2136d74