Fix unrestricted copying of backend TLS certs
With the change
https://review.opendev.org/c/openstack/kolla-ansible/+/915901
K-A copies backend TLS certificates to all containers of projects
when backend TLS for the project is enabled because of less strict
conditional check.

However, this causes K-A to attempt copying non-existent certificates
from hosts that is not part of backend TLS to containers that are
also not part of backend TLS.

This change introduces additional boolean variable
``copy_tls_backend`` that is used to check if we want to copy backend
TLS certificates to each container of projects.

If the variable is not set, the role ``service-cert-copy`` treats it
as false and does not copy backend TLS certificate to the container.

Cloeses-Bug: #2105505
Change-Id: I799dc6e6fdccc483784fc3c2088ada69f24412e9
(cherry picked from commit 52797d88d20d3e9be6574a9b81bead5c57a8552c)