Stop plugin installation via Grafana UI
Plugins must be installed during the Kolla image build. Installing
plugins via the UI causes Grafana backends to get out-of-sync in
multi-node deployments. In any deployment, any plugins installed
via the UI will be lost if the Grafana container is updated.

This change also prevents trying to fetch a plugin verification key
from grafana.com, since it is not required.

Change-Id: Iddc9b75b5c16bf44b38f2348b46a32b946b2428a
Closes-Bug: #2122587
Signed-off-by: Doug Szumski <doug@stackhpc.com>